+++ src/react/components/compositions/XUIGridAreaPrimary.d.ts
@@ -1,0 +1,7 @@
+import React from 'react';
+
+interface Props {
+  children?: React.ReactNode;
+}
+
+export default class XUIGridAreaPrimary extends React.PureComponent<Props> {}