+++ src/react/components/contentblock/XUIContentBlock.d.ts
@@ -1,0 +1,9 @@
+import React from 'react';
+
+interface Props {
+  children?: React.ReactNode;
+  className?: string;
+  qaHook?: string;
+}
+
+export default class XUIContentBlock extends React.PureComponent<Props> {}