+++ src/react/components/editabletable/XUIEditableTable.js
@@ -1,0 +1,9 @@
+import React, { PureComponent } from 'react';
+
+import { ns } from '../helpers/xuiClassNamespace';
+
+export default class XUIEditableTable extends PureComponent {
+  render() {
+    return <div className={`${ns}-editabletable`}>XUIEditableTable</div>;
+  }
+}