+++ src/react/components/modal/XUIModalBody.d.ts
@@ -1,0 +1,10 @@
+import React from 'react';
+
+interface Props {
+  children?: React.ReactNode;
+  className?: string;
+  qaHook?: string;
+}
+
+declare const XUIModalBody: React.FunctionComponent<Props>;
+export default XUIModalBody;