+++ src/react/components/structural/private/constants.ts
@@ -1,0 +1,12 @@
+export const rowVariants = {
+  flex: 'flex',
+  grid: 'grid',
+  standard: '',
+};
+
+export const columnShortNames = {
+  full: 12,
+  half: 6,
+  third: 4,
+  quarter: 3,
+};