+++ src/react/progressindicator.ts
@@ -1,6 +1,10 @@
 import XUIProgressCircular from './components/progressindicator/XUIProgressCircular';
 import XUIProgressLinear from './components/progressindicator/XUIProgressLinear';
 
-const progressTypes = { XUIProgressLinear, XUIProgressCircular };
+// Explicitly declaring the type here because of TS bug
+const progressTypes: {
+  XUIProgressCircular: typeof XUIProgressCircular;
+  XUIProgressLinear: typeof XUIProgressLinear;
+} = { XUIProgressLinear, XUIProgressCircular };
 
 export { progressTypes as default, XUIProgressLinear, XUIProgressCircular };